rbac: Move rbac package
This moves rbac package one level up in order share it
with the config package introduced in a later commit.
